test: added tests for DeserializeFromExcel, VerifyJsonPoolConfig

+++ Applications/ConsumerApi/test/ConsumerApi.Tests.Performance/tools/snapshot-creator-v2/Models/AppConfig.cs
@@ -6,5 +6,5 @@ namespace Backbone.ConsumerApi.Tests.Performance.SnapshotCreator.V2.Models;
 public record AppConfig : BaseConfig
 {
     [JsonPropertyName(nameof(TotalNumberOfRelationships))]
-    public int TotalNumberOfRelationships { get; set; }
+    public long TotalNumberOfRelationships { get; set; }
 }
